backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
anthropic==0.28.0
chromadb==0.5.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
pydantic==2.7.0
